requirements.txt
pytest
PyPDF2
faiss-gpu
sentence-transformers